fix(postgres): store full SchemaVersion range — BIGINT column, not INTEGER (#213)
The nixosTest ran and the shared assert_event_stream_conformance suite caught a
real bug: it feeds a boundary schema_version near u32::MAX, which is a valid
SchemaVersion (NonZeroU32) but exceeds i32::MAX, so the INTEGER column + reject
path refused a legitimate event. Widen the column to BIGINT; u32 -> i64 widens
infallibly, so the full range is stored with no rejection.

Concretely: schema_version column INTEGER->BIGINT; EventRow/PreparedInsert
schema_version i32->i64; prepare_inserts drops the i32::try_from + CorruptRow
arm for an infallible i64::from; the unit test flips from
"over-i32-max-is-error" to "full-u32-range-accepted". Plan + findings updated —
this is a concrete #213 win: the 2nd adapter + conformance suite together caught
a silent data-rejection bug fjall never exposed.

30/31 nixosTest tests already passed (incl. all $all no-skip proofs); this is
the last one.

+++ crates/nexus-postgres/src/schema.rs
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS events (
     version        BIGINT   NOT NULL,
     txid           xid8     NOT NULL DEFAULT pg_current_xact_id(),
     event_type     TEXT     NOT NULL,
-    schema_version INTEGER  NOT NULL,
+    schema_version BIGINT   NOT NULL,
     payload        BYTEA    NOT NULL,
     metadata       BYTEA,
     PRIMARY KEY (global_seq),
